requirements.txt
automata-lib==5.0.0
pytest